+++ CHANGELOG.md
@@ -8,7 +8,7 @@
   - `StateExecution.getAllChildWorkflows` does not return current state variables of child workflows anymore.
   - `WorkflowStateProcessor` does not read child workflow IDs of processed workflow instances anymore.
   - `QueryWorkflowInstances.Builder` does not query for current state variables by default anymore.
-  - `BulkWorklow` moved to `curated` java package.
+  - `BulkWorklow` moved to `io.nflow.engine.workflow.curated` java package.
 
 **Highlights**
 - `nflow-engine`
@@ -36,6 +36,7 @@
   - To get child workflows with state variables in a state method, use `StateExecution.queryChildWorkflows`.
   - To get all child workflow IDs in a state method, use `StateExecution.getAllChildWorkflows`.
   - New `MaintenanceWorkflow` that can be enabled by setting `nflow.maintenance.insertWorkflowIfMissing` property to true. On the first run the workflow is inserted and it will start to run periodically. Further configuration must be done by editing the two state variables of the workflow: `cron` for the scheduling and `config` for the `MaintenanceConfiguration` to execute.
+    - The first time configuration can also be tuned with `nflow.maintenance.initial.cron`, `nflow.maintenance.initial.delete.olderThan`, `nflow.maintenance.initial.archive.olderThan` and `nflow.maintenance.initial.deleteArchived.olderThan`
 - `nflow-explorer`
   - Dependency updates:
     - angular extra libraries 1.7.9